refactor: shared/components
