+++ frontend/src/app/layout.tsx
@@ -1,8 +1,8 @@
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import Header from './shared/header'
-import { Children } from './shared/types'
+import { Children } from '../shared/components/types'
+import Header from '../shared/components/header'
 
 const inter = Inter({ subsets: ['latin'] })
